refactor(ricer::config)!: add `Entry` trait
BREAKING CHANGE: unify intermediary TOML types under one interface.
