+++ src/config/entry.rs
@@ -2,8 +2,13 @@
 // SPDX-License-Identifier: MIT
 
 use std::fmt;
-use toml_edit::visit::{visit_table_like_kv, Visit};
-use toml_edit::{Array, Item, Key, Table, Value};
+use toml_edit::visit::{visit_inline_table, visit_table_like_kv, Visit};
+use toml_edit::{Array, Item, Key, Table, Value, InlineTable};
+use std::path::PathBuf;
+
+pub trait Entry<'toml> {
+    fn to_toml(&self) -> (Key, Item);
+}
 
 /// Repository configuration settings.
 ///
@@ -59,7 +64,10 @@ impl Repository {
         self
     }
 
-    pub fn to_toml(&self) -> (Key, Item) {
+}
+
+impl Entry<'_> for Repository {
+    fn to_toml(&self) -> (Key, Item) {
         let mut repo = Table::new();
         let mut repo_bootstrap = Table::new();
 
@@ -266,5 +274,136 @@ impl fmt::Display for OsType {
             OsType::MacOs => write!(f, "macos"),
             OsType::Windows => write!(f, "windows"),
         }
+    }
+}
+
+/// Command hook settings.
+///
+/// An intermediary structure to help deserialize and serialize command hook
+/// from Ricer's command hook configuration file.
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub struct CommandHook {
+    /// Name of command to bind hook definitions too.
+    pub cmd: String,
+
+    /// Array of hook definitions to execute.
+    pub hooks: Vec<Hook>,
+}
+
+impl CommandHook {
+    pub fn new(cmd: impl Into<String>) -> Self {
+        Self { cmd: cmd.into(), hooks: Default::default() }
+    }
+
+    pub fn add_hook(mut self, hook: Hook) -> Self {
+        self.hooks.push(hook);
+        self
+    }
+}
+
+impl Entry<'_> for CommandHook {
+    fn to_toml(&self) -> (Key, Item) {
+        let mut tables = Array::new();
+        let mut iter = self.hooks.iter().enumerate().peekable();
+        while let Some((_, hook)) = iter.next() {
+            let mut inline = InlineTable::new();
+            let decor = inline.decor_mut();
+            decor.set_prefix("\n    ");
+
+            if iter.peek().is_none() {
+                decor.set_suffix("\n");
+            }
+
+            if let Some(pre) = &hook.pre {
+                inline.insert("pre", Value::from(pre));
+            }
+
+            if let Some(post) = &hook.post {
+                inline.insert("post", Value::from(post));
+            }
+
+            if let Some(workdir) = &hook.workdir {
+                inline.insert("workdir", Value::from(String::from(workdir.to_string_lossy())));
+            }
+
+            tables.push_formatted(Value::from(inline));
+        }
+
+        let key = Key::new(&self.cmd);
+        let value = Item::Value(Value::from(tables));
+        (key, value)
+    }
+}
+
+fn from_toml<'toml>(entry: (&'toml Key, &'toml Item)) -> CommandHook {
+    let (key, value) = entry;
+    let mut cmd_hook = CommandHook::new(key.get());
+    cmd_hook.visit_item(value);
+    cmd_hook
+}
+
+impl<'toml> From<(&'toml Key, &'toml Item)> for CommandHook {
+    fn from(entry: (&'toml Key, &'toml Item)) -> Self {
+        from_toml(entry)
+    }
+}
+
+impl From<(Key, Item)> for CommandHook {
+    fn from(entry: (Key, Item)) -> Self {
+        let (key, value) = entry;
+        from_toml((&key, &value))
+    }
+}
+
+impl<'toml> Visit<'toml> for CommandHook {
+    fn visit_inline_table(&mut self, node: &'toml InlineTable) {
+        let pre = if let Some(pre) = node.get("pre") { pre.as_str() } else { None };
+        let post = if let Some(post) = node.get("post") { post.as_str() } else { None };
+        let workdir = if let Some(workdir) = node.get("workdir") { workdir.as_str() } else { None };
+
+        let hook = Hook::new();
+        let hook = if let Some(pre) = pre { hook.pre(pre) } else { hook };
+        let hook = if let Some(post) = post { hook.post(post) } else { hook };
+        let hook = if let Some(workdir) = workdir { hook.workdir(workdir) } else { hook };
+        self.hooks.push(hook);
+
+        visit_inline_table(self, node);
+    }
+}
+
+/// Hook definition settings.
+///
+/// An intermediary structure to help deserialize and serialize hook entries
+/// for command hook settings in command hook configuration file.
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub struct Hook {
+    /// Execute hook script _before_ command itself.
+    pub pre: Option<String>,
+
+    /// Execute hook script _after_ command itself.
+    pub post: Option<String>,
+
+    /// Set working directory of hook script.
+    pub workdir: Option<PathBuf>,
+}
+
+impl Hook {
+    pub fn new() -> Self {
+        Default::default()
+    }
+
+    pub fn pre(mut self, script: impl Into<String>) -> Self {
+        self.pre = Some(script.into());
+        self
+    }
+
+    pub fn post(mut self, script: impl Into<String>) -> Self {
+        self.post = Some(script.into());
+        self
+    }
+
+    pub fn workdir(mut self, path: impl Into<PathBuf>) -> Self {
+        self.workdir = Some(path.into());
+        self
     }
 }
